libs, vm: Move the 'add_volume_disk' helper to the 'vm' module
The helper is generic, fits better closer to the vm module and has now
another usage there.
